General python3 support
- Better py3 support removing imp when possible
- Default message personalization
- Possibility of creating a custom docs reply
- PluginManagerinto it's own file as is used by dispatcher too
