+++ README.md
@@ -27,9 +27,13 @@ First you need to get the slack api token for your bot. You have two options:
 1. If you use a [bot user integration](https://api.slack.com/bot-users) of slack, you can get the api token on the integration page.
 2. If you use a real slack user, you can generate an api token on [slack web api page](https://api.slack.com/web).
 
-### Configure the api token
 
-Then you need to configure the `API_TOKEN` in a python module `slackbot_settings.py`, which must be located in a python import path.
+### Configure the bot
+First create a `slackbot_settings.py` and a `run.py` in your own instance of slackbot.
+
+##### Configure the api token
+
+Then you need to configure the `API_TOKEN` in a python module `slackbot_settings.py`, which must be located in a python import path. This will be automatically imported by the bot.
 
 slackbot_settings.py:
 
@@ -39,7 +43,7 @@ API_TOKEN = "<your-api-token>"
 
 Alternatively, you can use the environment variable `SLACKBOT_API_TOKEN`.
 
-### Run the bot
+##### Run the bot
 
 ```python
 from slackbot.bot import Bot
@@ -49,6 +53,24 @@ def main():
 
 if __name__ == "__main__":
     main()
+```
+##### Configure the default answer
+Add to `slackbot_settings.py` a default_reply:
+```python
+default_reply = "Sorry but I didn't understood you" 
+```
+
+##### Configure the docs answer
+The `message` attribute passed to [your custom plugins](#create-plugins) has an special function `message.docs_reply()` that will parse all the plugins available and return the Docs in each of them.
+
+##### Configure the plugins
+Add [your plugin modules](#create-plugins) to a `PLUGINS` list in `slackbot_settings.py`:
+
+```python
+PLUGINS = [
+    'slackbot.plugins',
+    'mybot.plugins',
+]
 ```
 
 Now you can talk to your bot in your slack client!
@@ -73,8 +95,7 @@ def github():
     }]
     message.send_webapi('', json.dumps(attachments))
 ```
-
-## Plugins
+## Create Plugins
 
 A chat bot is meaningless unless you can extend/customize it to fit your own use cases.
