+++ slackbot/bot.py
@@ -3,15 +3,13 @@ from __future__ import absolute_import
 import imp
 import importlib
 import logging
-import os
 import re
 import time
 from glob import glob
 from six.moves import _thread
-
 from slackbot import settings
+from slackbot.manager import PluginsManager
 from slackbot.slackclient import SlackClient
-from slackbot.utils import to_utf8
 from slackbot.dispatcher import MessageDispatcher
 
 logger = logging.getLogger(__name__)
@@ -21,8 +19,8 @@ class Bot(object):
     def __init__(self):
         self._client = SlackClient(
             settings.API_TOKEN,
-            bot_icon = settings.BOT_ICON if hasattr(settings, 'BOT_ICON') else None,
-            bot_emoji = settings.BOT_EMOJI if hasattr(settings, 'BOT_EMOJI') else None
+            bot_icon=settings.BOT_ICON if hasattr(settings, 'BOT_ICON') else None,
+            bot_emoji=settings.BOT_EMOJI if hasattr(settings, 'BOT_EMOJI') else None
         )
         self._plugins = PluginsManager()
         self._dispatcher = MessageDispatcher(self._client, self._plugins)
@@ -40,50 +38,6 @@ class Bot(object):
         while True:
             time.sleep(30 * 60)
             self._client.ping()
-
-
-class PluginsManager(object):
-    commands = {
-        'respond_to': {},
-        'listen_to': {}
-    }
-
-    def __init__(self):
-        pass
-
-    def init_plugins(self):
-        if hasattr(settings, 'PLUGINS'):
-            plugins = settings.PLUGINS
-        else:
-            plugins = 'slackbot.plugins'
-
-        for plugin in plugins:
-            self._load_plugins(plugin)
-
-    def _load_plugins(self, plugin):
-        logger.info('loading plugin "%s"', plugin)
-        path_name = None
-        for mod in plugin.split('.'):
-            if path_name is not None:
-                path_name = [path_name]
-            _, path_name, _ = imp.find_module(mod, path_name)
-        for pyfile in glob('{}/[!_]*.py'.format(path_name)):
-            module = '.'.join((plugin, os.path.split(pyfile)[-1][:-3]))
-            try:
-                importlib.import_module(module)
-            except:
-                logger.exception('Failed to import %s', module)
-
-    def get_plugins(self, category, text):
-        has_matching_plugin = False
-        for matcher in self.commands[category]:
-            m = matcher.search(text)
-            if m:
-                has_matching_plugin = True
-                yield self.commands[category][matcher], to_utf8(m.groups())
-
-        if not has_matching_plugin:
-            yield None, None
 
 
 def respond_to(matchstr, flags=0):
