+++ slackbot/dispatcher.py
@@ -6,6 +6,7 @@ import re
 import time
 import traceback
 from six import iteritems
+from slackbot.manager import PluginsManager
 from slackbot.utils import to_utf8, WorkerPool
 
 logger = logging.getLogger(__name__)
@@ -96,20 +97,28 @@ class MessageDispatcher(object):
             time.sleep(1)
 
     def _default_reply(self, msg):
-        default_reply = [
-            u'Bad command "%s", You can ask me one of the following questions:\n' % msg['text'],
-        ]
-        default_reply += [u'    • `{0}` {1}'.format(p.pattern, v.__doc__ or "")
-                          for p, v in iteritems(self._plugins.commands['respond_to'])]
-            
-        self._client.rtm_send_message(msg['channel'],
-                                      '\n'.join(to_utf8(default_reply)))
+        try:
+            from slackbot_settings import default_reply
+            default_reply = to_utf8(default_reply)
+
+        except ImportError:
+
+            default_reply = [
+                'Bad command "%s", You can ask me one of the following questions:\n' % msg['text'],
+            ]
+            default_reply += ['    • `{0}` {1}'.format(p.pattern, v.__doc__ or "")
+                              for p, v in iteritems(self._plugins.commands['respond_to'])]
+
+            default_reply = '\n'.join(to_utf8(default_reply))
+
+        self._client.rtm_send_message(msg['channel'], default_reply)
 
 
 class Message(object):
     def __init__(self, slackclient, body):
         self._client = slackclient
         self._body = body
+        self._plugins = PluginsManager()
 
     def _get_user_id(self):
         if 'user' in self._body:
@@ -118,7 +127,7 @@ class Message(object):
         return self._client.find_user_by_name(self._body['username'])
 
     def _gen_at_message(self, text):
-        text = u'<@{}>: {}'.format(self._get_user_id(), text)
+        text = '<@{}>: {}'.format(self._get_user_id(), text)
         return text
 
     def _gen_reply(self, text):
@@ -186,3 +195,7 @@ class Message(object):
     @property
     def body(self):
         return self._body
+
+    def docs_reply(self):
+        reply = ['    • `{0}` {1}'.format(v.__name__, v.__doc__ or "") for p, v in iteritems(self._plugins.commands['respond_to'])]
+        return '\n'.join(to_utf8(reply))
