+++ slackbot/plugins/upload.py
@@ -2,6 +2,7 @@ import os
 from slackbot.bot import respond_to
 from slackbot.utils import download_file, create_tmp_file
 
+
 @respond_to(r'upload \<?(.*)\>?')
 def upload(message, url):
     url = url.lstrip('<').rstrip('>')
