+++ slackbot/settings.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 import os
 
 DEBUG = False
