+++ slackbot/slackclient.py
@@ -1,4 +1,4 @@
-#coding: UTF-8
+# -*- coding: utf-8 -*-
 
 from __future__ import print_function, absolute_import
 import os
