+++ slackbot/utils.py
@@ -1,5 +1,4 @@
-#coding: UTF-8
-
+# -*- coding: utf-8 -*-
 
 import os
 import logging
